Response to code review
