+++ src/phoebus_utils/programming_utils.hpp
@@ -27,18 +27,12 @@ struct is_specialization_of<TEMPL<TPARAMS...>, TEMPL> : public std::true_type {}
 
 template <typename T>
 bool soft_equiv(const T x, const T y, const T tolerance = static_cast<T>(1.e-10f)) {
-  bool equivalent = false;
-  if (2. * fabs(y - x) < tolerance * (std::fabs(x) + std::fabs(y))) {
-    equivalent = true;
-  }
-
-  return equivalent;
+  return (2. * fabs(y - x) < tolerance * (std::fabs(x) + std::fabs(y)));
 }
 
 template <typename T>
 bool ContainedInVector(T &item, std::vector<T> &container) {
   if (std::find(container.begin(), container.end(), item) == container.end()) {
-    std::stringstream msg;
     return false;
   }
   return true;
