+++ src/radiation/closure_mocmc.hpp
@@ -53,7 +53,7 @@ class ClosureMOCMC : public ClosureEdd<Vec, Tens2, SET> {
     PARTHENON_FAIL("MOCMC has already stored TilPi elsewhere!");
 
     // Do nothing
-    return ClosureStatus::success;
+    return ClosureStatus::fail;
   }
 
   KOKKOS_FUNCTION
@@ -62,7 +62,7 @@ class ClosureMOCMC : public ClosureEdd<Vec, Tens2, SET> {
     PARTHENON_FAIL("MOCMC has already stored TilPi elsewhere!");
 
     // Do nothing
-    return ClosureStatus::success;
+    return ClosureStatus::fail;
   }
 };
 } // namespace radiation
